[refactor][FIB-560] move the coincidence viewer into the application's ui package
The coincidence viewer's only host is AutoLamella: AutoLamellaUI opens it from a
dev-menu action, and nothing else constructs it except its own main() harness. It
takes an Experiment and selects a Lamella. By every observable measure it is an
AutoLamella widget sitting in the generic package, so it moves, as the autolamella_*
and lamella/workflow groups did before it.

The widget itself is byte-identical: its imports of LamellaNameListWidget and
SelectedLamellaWidget already pointed into the application package after FIB-559,
and are now intra-package.

coincidence_milling_confirmation_dialog.py deliberately stays under fibsem/ui. It
has no AutoLamella imports, so it is genuinely generic, and an application widget
importing a generic one is the right direction.

fibsem/ui modules importing AutoLamella drops from 2 to 1. The remaining one is
FibsemMinimapWidget, which cannot move -- FibsemUI is a live generic host for it --
and whose decoupling is recorded on FIB-405, the phase that rewrites that file
anyway.

Also repairs a coverage regression that this move made visible but did not cause
alone. tests/ui/test_shared_palette.py anchors its walk at fibsem/ui, so relocating
25 widgets across FIB-558/559/560 dropped 22 files out of the palette guard -- 11 of
them real widgets. The test kept passing with less to test; only a collected-node
diff against origin/main showed it. It now matches on a `ui` path component, as
tests/test_ui_extra_isolation.py already did, taking coverage from 192 to 213
modules with no new violations. That function's docstring had already recorded two
earlier versions of the same mistake; this is the third.
