+++ fibsem/applications/autolamella/ui/fluorescence_coincidence_viewer_widget.py
