+++ tests/ui/test_gui_thread_migration.py
@@ -45,7 +45,7 @@ MIGRATED_MODULES = [
     "fibsem/ui/FibsemMinimapWidget.py",
     "fibsem/ui/widgets/milling_widget.py",
     "fibsem/ui/widgets/fluorescence_control_widget.py",
-    "fibsem/ui/widgets/fluorescence_coincidence_viewer_widget.py",
+    "fibsem/applications/autolamella/ui/fluorescence_coincidence_viewer_widget.py",
     "fibsem/applications/autolamella/ui/AutoLamellaUI.py",
 ]
